Slightly more aggressive detangling, Five endpoints now added to random networks

+++ src/dream/team/assemble/gui/Button.java
@@ -25,7 +25,7 @@ public class Button
         }
     };
 
-    public Button(int x, int y, int width, int height, String label)
+    public Button(float x, float y, float width, float height, String label)
     {
         this.x = x;
         this.y = y;
